v4: support for SO_BINDTODEVICE
In case the hostns already has a device with the service address assigned
to it, the health checker should use SO_BINDTODEVICE to pin it to the phys
dev instead.

Example described below:

On LB node:

  root@apoc:~# ip a
  [...]
  6: enp2s0np0: <BROADCAST,MULTICAST,UP,LOWER_UP> mtu 1500 xdp/id:1402 qdisc mq state UP [...]  <--- XDP LB dev
    link/ether 00:15:4d:13:04:dc brd ff:ff:ff:ff:ff:ff
    inet 192.168.180.29/24 scope global enp2s0np0
       valid_lft forever preferred_lft forever
    inet6 f00d::29/24 scope global
       valid_lft forever preferred_lft forever
    inet6 fe80::215:4dff:fe13:4dc/64 scope link
       valid_lft forever preferred_lft forever
  [...]
  9: cilium_ipip4@NONE: <NOARP,UP,LOWER_UP> mtu 1480 qdisc noqueue state UNKNOWN group default qlen 1000
    link/ipip 0.0.0.0 brd 0.0.0.0
  10: dummy0: <BROADCAST,NOARP,UP,LOWER_UP> mtu 1500 qdisc noqueue state UNKNOWN group default qlen 1000
    link/ether 7e:1a:64:82:bd:61 brd ff:ff:ff:ff:ff:ff
    inet 7.7.7.7/32 scope global dummy0                      <--- SVC addr
       valid_lft forever preferred_lft forever
    inet6 fe80::7c1a:64ff:fe82:bd61/64 scope link
       valid_lft forever preferred_lft forever

Without SO_BINDTODEVICE, the request does /not/ go out the enp2s0np0 dev
where the tc egress BPF prog is sitting that will handle & push the tunnel
case. Instead, the traffic is looped back via lo and we get a RST:

  root@apoc:~# tcpdump -i lo -n host 7.7.7.7 or host 192.168.180.28
  [...]
  13:56:53.628948 IP 7.7.7.7.41665 > 7.7.7.7.8080: Flags [S], seq 93515663, win 65495, options [mss 65495,sackOK,TS val 2930611131 ecr 0,nop,wscale 7], length 0
  13:56:53.628957 IP 7.7.7.7.8080 > 7.7.7.7.41665: Flags [R.], seq 0, ack 93515664, win 0, length 0

With this patch, we pin the dev to enp2s0np0 via SO_BINDTODEVICE and execute
the request again on the same env as above:

  root@apoc:~# ./v4 enp2s0np0 7.7.7.7 8080 192.168.180.28

Now the request goes out the node and the remote host can see it:

  root@tank:~# tcpdump -i enp2s0np0 -n
  [...]
  14:07:06.767306 IP 192.168.180.29 > 192.168.180.28: IP 192.168.180.29.35697 > 7.7.7.7.8080: Flags [S], seq 225468299, win 64240, options [mss 1460,sackOK,TS val 3042231612 ecr 0,nop,wscale 7], length 0 (ipip-proto-4)

For testing the old behavior, we also support the "any" keyword where the
SO_BINDTODEVICE is not performed:

  root@apoc:~# ./v4 any 7.7.7.7 8080 192.168.180.28

Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>